Move SimpleLongSerializer to its own file
